feat/ 심플 MCP서버와 , device_test.py 바꿨다 trigger.py였던 MCP서버 심플 비밥 MCP로 바꿨고 디바이스 테스트 .py 안의 내용도 간단하게 솔로기능만 구현할수있도록 줄였다
